2차 수정

+++ src/main/java/com/bitc/project/service/MemberServiceImpl.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import com.bitc.project.dao.MemberDAO;
+import com.bitc.project.util.Criteria;
+import com.bitc.project.util.PageMaker;
 import com.bitc.project.vo.LoginDTO;
 import com.bitc.project.vo.MemberVO;
 
@@ -43,6 +45,11 @@ public class MemberServiceImpl implements MemberService {
 
     @Override
     public void deleteMember(String ID) {
+    	
+    	if (ID == "ADMIN" && ID == "admin") {
+            throw new RuntimeException("관리자 계정은 삭제할 수 없습니다.");
+        }
+    	
         memberDAO.deleteMember(ID);
     }
 
@@ -60,6 +67,11 @@ public class MemberServiceImpl implements MemberService {
 	public String checkNick(String nname) {
 		return memberDAO.checkNick(nname) == null ? "OK" : "X";
 	}
+
+	/*
+	 * @Override public PageMaker getPageMaker(Criteria cri, List<MemberVO> members)
+	 * { return new PageMaker(cri,memberDAO.totalCount(members)); }
+	 */
